+++ mods/CreateDragonPlus/src/main/java/plus/dragons/createdragonsplus/config/CDPFeaturesConfig.java
@@ -30,6 +30,10 @@ public class CDPFeaturesConfig extends FeaturesConfig {
             true,
             "fluid/dragon_breath",
             Comments.dragonBreathFluid).addAlias("dragons_breath_fluid");
+    public final ConfigFeature dyeFluidsLavaInteractionGenerateColoredConcrete = feature(
+            true,
+            "fluid/dye/lava_interaction_generate_colored_concrete",
+            Comments.dyeFluidsLavaInteractionGenerateColoredConcrete);
     private final ConfigGroup block = group(1, "block", Comments.block);
     public final ConfigFeature fluidHatch = feature(
             true,
@@ -65,6 +69,7 @@ public class CDPFeaturesConfig extends FeaturesConfig {
     static class Comments {
         static final String fluid = "Fluid Feature Elements";
         static final String dyeFluids = "If Dye Fluids should be enabled";
+        static final String dyeFluidsLavaInteractionGenerateColoredConcrete = "If colored concrete should be generated when Fye Fluid meets Lava";
         static final String dragonBreathFluid = "If Dragon's Breath Fluid should be enabled";
         static final String block = "Block Feature Elements";
         static final String fluidHatch = "If Fluid Hatch should be enabled";